# perf: optimize URLPattern