```
$ deno run urlpattern_bench_exec.js
took 2037 ms
$ deno run urlpattern_bench_test.js
took 2004 ms

$ ./target/release/deno run urlpattern_bench_exec.js
took 1727 ms
$ ./target/release/deno run urlpattern_bench_test.js
took 1101 ms
```

That’s about 15% speed up for exec and about 40% speed up for test.
I didn’t do anything magical, just switched from serializing structs (ie. JS objects) to serializing arrays (and turning them into objects in JS if needed).